+++ spec/support/servers/plugins/login_script.rb
@@ -1,0 +1,13 @@
+require 'sinatra'
+require 'sinatra/contrib'
+
+get '/' do
+    cookies[:success] ||= 'false'
+
+    if cookies[:success] == 'true'
+        <<-HTML
+            <a href='/congrats'>Hi there logged-in user!</a>
+        HTML
+    end
+end
+